cleanup and consolidation
